Improved error handling and queue manipulation.
Handle errors from the submission process by retrying every 10s for up to a
minute and then switching to offline mode and popping up a dialog instructing
the user to use the preferences to go back online.

Gate the submission loop on not being broken or offline.

Only remove a listen from the queue if it was successfully sent instead of
messing around reinserting on error.